requirements.txt
Flask==3.0.2
google-cloud-storage==2.9.0
google-cloud-secret-manager==2.16.2
google-generativeai==0.8.4
gunicorn==21.2.0
Flask-Cors==4.0.0
pillow